Prevent layout shift
Some of the "words" in the table are so long they cause the table to grow beyond its specified width (i.e. 100% of its parent width).  Worse, because of the way the table data is rendered one row at a time, it creates a janky effect where the table gets wider as it loads.

This could be (probably is?) because of some of the junk in the database that needs to be fixed, but in any event, this little bit of CSS improves the situation a lot.