Add a super-simple click-to-show/hide functionality
The principle here is to hide all the rows with the class `extra-details` and then toggle showing/hiding them when the main row for that db record (the immediately preceding one) is clicked.  There are lots of ways this could be done, but this is the simplest, and it's effective.